[Added] OC-730  #comment Fixed bug where type of schedulerId from token claims was Long #time 10m

+++ src/backend/src/main/java/com/becon/opencelium/backend/mysql/service/WebhookServiceImp.java
@@ -61,7 +61,7 @@ public class WebhookServiceImp implements WebhookService {
 
         String uuid = claims.getClaim("uuid").toString();
         String tmp = claims.getClaim("userId").toString();
-        int schedulerId =  (int) claims.getClaim("schedulerId");
+        int schedulerId = ((Long)claims.getClaim("schedulerId")).intValue();
         int userId = Integer.parseInt(tmp);
 
         boolean userExists = userService.existsById(userId);
